experiment page visual tweaks
